Refactoring: creating folder structure, relocating TodoContext to Data Folder
