+++ Controllers/TodoItemsController.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using TodoApi.Models;
+using TodoApiDTO.Data;
 
 namespace TodoApi.Controllers
 {
