+++ Data/TodoContext.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using TodoApi.Models;
 
-namespace TodoApi.Models
+namespace TodoApiDTO.Data
 {
     public class TodoContext : DbContext
     {
