+++ Services/Interfaces/ITodoItemService.cs
@@ -1,0 +1,6 @@
+﻿namespace TodoApiDTO.Services.Interfaces
+{
+    public interface ITodoItemService
+    {
+    }
+}
